+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Mempool/API.hs
@@ -30,12 +30,6 @@ module Ouroboros.Consensus.Mempool.API (
   , TicketNo
   , TxSizeInBytes
   , zeroTicketNo
-    -- * Deprecated re-exports
-  , MempoolCapacityBytes
-  , MempoolCapacityBytesOverride
-  , MempoolSize
-  , TraceEventMempool
-  , computeMempoolCapacity
   ) where
 
 import           Ouroboros.Consensus.Block (SlotNo)
@@ -362,27 +356,3 @@ data MempoolSnapshot blk = MempoolSnapshot {
     -- | The ledger state after all transactions in the snapshot
   , snapshotLedgerState :: TickedLedgerState blk
   }
-
-{-------------------------------------------------------------------------------
-  Deprecations
--------------------------------------------------------------------------------}
-
-{-# DEPRECATED MempoolCapacityBytes "Use Ouroboros.Consensus.Mempool (MempoolCapacityBytes)" #-}
-type MempoolCapacityBytes = Cap.MempoolCapacityBytes
-
-{-# DEPRECATED MempoolSize "Use Ouroboros.Consensus.Mempool (MempoolSize)" #-}
-type MempoolSize = Cap.MempoolSize
-
-{-# DEPRECATED MempoolCapacityBytesOverride "Use Ouroboros.Consensus.Mempool (MempoolCapacityBytesOverride)" #-}
-type MempoolCapacityBytesOverride = Cap.MempoolCapacityBytesOverride
-
-{-# DEPRECATED computeMempoolCapacity "Use Ouroboros.Consensus.Mempool (computeMempoolCapacity)" #-}
-computeMempoolCapacity
-  :: LedgerSupportsMempool blk
-  => TickedLedgerState blk
-  -> MempoolCapacityBytesOverride
-  -> MempoolCapacityBytes
-computeMempoolCapacity = Cap.computeMempoolCapacity
-
-{-# DEPRECATED TraceEventMempool "Use Ouroboros.Consensus.Mempool (TraceEventMempool)" #-}
-data TraceEventMempool