+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Mempool/Update.hs
@@ -7,9 +7,6 @@ module Ouroboros.Consensus.Mempool.Update (
     implAddTx
   , implRemoveTxs
   , implSyncWithLedger
-    -- * Exported for deprecated modules
-  , pureRemoveTxs
-  , pureSyncWithLedger
   ) where
 
 import           Control.Concurrent.Class.MonadMVar (MVar, MonadMVar, withMVar)
@@ -20,9 +17,7 @@ import qualified Data.Set as Set
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.SupportsMempool
-import           Ouroboros.Consensus.Mempool.API hiding (MempoolCapacityBytes,
-                     MempoolCapacityBytesOverride, MempoolSize,
-                     TraceEventMempool, computeMempoolCapacity)
+import           Ouroboros.Consensus.Mempool.API
 import           Ouroboros.Consensus.Mempool.Capacity
 import           Ouroboros.Consensus.Mempool.Impl.Common
 import           Ouroboros.Consensus.Mempool.TxSeq (TxTicket (..))