+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Storage/ChainDB/Impl/Types.hs
@@ -492,7 +492,6 @@ getBlockToAdd (BlocksToAdd queue) = atomically $ readTBQueue queue
   Trace types
 -------------------------------------------------------------------------------}
 
-{-# DEPRECATED TraceLedgerEvent "Use TraceSnapshotEvent"#-}
 -- | Trace type for the various events of the ChainDB.
 data TraceEvent blk
   = TraceAddBlockEvent          (TraceAddBlockEvent           blk)
@@ -503,7 +502,6 @@ data TraceEvent blk
   | TraceOpenEvent              (TraceOpenEvent               blk)
   | TraceIteratorEvent          (TraceIteratorEvent           blk)
   | TraceSnapshotEvent          (LgrDB.TraceSnapshotEvent     blk)
-  | TraceLedgerEvent            (LgrDB.TraceSnapshotEvent     blk)
   | TraceLedgerReplayEvent      (LgrDB.TraceReplayEvent       blk)
   | TraceImmutableDBEvent       (ImmutableDB.TraceEvent       blk)
   | TraceVolatileDBEvent        (VolatileDB.TraceEvent        blk)